Revert "klassnamn = tabellnamn (små bokstäver)"
This reverts commit 110d447f

+++ src/main/java/com/example/Walkocity/accessingdatamysql/MainController.java
@@ -106,7 +106,7 @@ public class MainController {
         // @ResponseBody means the returned String is the response, not a view name
         // @RequestParam means it is a parameter from the GET or POST request
         WalkEvent walkEvent = new WalkEvent();
-        walkEvent.setWalkEventId(Integer.parseInt(walkEventId));
+        walkEvent.setWalkEventId(Long.parseLong(walkEventId));
         walkEvent.setOwnerId(Integer.parseInt(ownerId));
         walkEvent.setLocation(location);
         walkEvent.setDate(date);
