+++ src/main/java/com/example/Walkocity/accessingdatamysql/WalkEvent.java
@@ -8,7 +8,7 @@ public class WalkEvent implements Serializable { //fundera över hur vilken klas
     private String eventName;
 
     @Id @Column(updatable = false)
-    private Integer walkEventId;
+    private Long walkEventId;
     @Id @Column(updatable = false)
     private Integer ownerId; //Ersätts av User-instans eller liknande
     @Id @Column(updatable = false)
@@ -29,11 +29,11 @@ public class WalkEvent implements Serializable { //fundera över hur vilken klas
         return eventName;
     }
 
-    public Integer getWalkEventId() {
+    public Long getWalkEventId() {
         return walkEventId;
     }
 
-    public void setWalkEventId(Integer walkEventId) {
+    public void setWalkEventId(Long walkEventId) {
         this.walkEventId = walkEventId;
     }
